Added User & Base Class models.
Not sure if its the right thing to do but i added the custom Django
User & UserManager models. I thought that if at anypoint we wanted to
add fields to the users this would be the best way to do it. I also
used the 'Base' class, using UUIDField might be better, but i was
thinking if we need to override any methods at any point, this is the
place to do it.